# backend_go/docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  models.History:
    properties:
      date:
        type: string
      id:
        type: integer
      nb_jetons:
        type: integer
      stand_name:
        type: string
      user_id:
        type: integer
    type: object
...
  models.Kermesse:
    properties:
      id:
        type: integer
      name:
        type: string
      organisateurs:
        description: 'Relations Many-to-Many : Organisateurs et participants de la
          kermesse'
        items:
          $ref: '#/definitions/models.User'
        type: array
      participants:
        items:
          $ref: '#/definitions/models.User'
        type: array
      picture:
        type: string
      stands:
        items:
          $ref: '#/definitions/models.Stand'
        type: array
      user_id:
        description: 'Relation Many-to-One : L''utilisateur qui crée la kermesse'
        type: integer
    type: object
  models.Product:
    properties:
      id:
        type: integer
      jetons_requis:
        type: integer
      name:
        type: string
      nb_products:
        type: integer
      picture:
        type: string
      stand_id:
        description: Clé étrangère vers le stand
        type: integer
      type:
        type: string
    type: object
  models.Stand:
    properties:
      conso:
        type: integer
      id:
        type: integer
      jetons_requis:
        type: integer
      kermesses:
        items:
          $ref: '#/definitions/models.Kermesse'
        type: array
      name:
        type: string
      pts_donnees:
        type: integer
      stocks:
        description: Clé étrangère vers Product
        items:
          $ref: '#/definitions/models.Product'
        type: array
      type:
        type: string
      user_id:
        type: integer
    type: object
  models.Transaction:
    properties:
      Quantity:
        type: integer
      date_transaction:
        type: string
      id:
        type: integer
      price:
        type: number
      type:
        type: string
      user_id:
        description: Relations avec l'utilisateur
        type: integer
    type: object
  models.User:
    properties:
      email:
        type: string
      enfants:
        items:
          $ref: '#/definitions/models.User'
        type: array
      firstname:
        type: string
      historique:
        items:
          $ref: '#/definitions/models.History'
        type: array
      id:
        type: integer
      jetons:
        type: integer
      kermesses:
        description: Relations One-to-Many pour les Kermesses et Stands gérés par
          l'utilisateur
        items:
          $ref: '#/definitions/models.Kermesse'
        type: array
      lastname:
        type: string
      parents:
        description: Relations Many-to-Many pour Parents/Enfants
        items:
          $ref: '#/definitions/models.User'
        type: array
      password:
        type: string
      picture:
        type: string
      pts_attribues:
        type: integer
      role:
        description: 1 = ADMIN / 2 = ORGANISATEUR / 3 = TENEUR DE STAND / 4 = PARENT
          / 5 ELEVE
        type: integer
      stands:
        items:
          $ref: '#/definitions/models.Stand'
        type: array
      transactions:
        items:
          $ref: '#/definitions/models.Transaction'
        type: array
    type: object
...
  requests.AddStandRequest:
    properties:
      stand_ids:
        items:
          type: integer
        type: array
    required:
    - stand_ids
    type: object
  requests.AddUserRequest:
    properties:
      type:
        type: string
      user_ids:
        items:
          type: integer
        type: array
    type: object
...
  requests.KermeseRequest:
    properties:
      name:
        type: string
      picture:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /create-kermesse:
    post:
      consumes:
      - application/json
      description: Permet aux user de créé un groupe de groupeVoyage
      parameters:
      - default: Bearer Add access token here
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Données de la kermesse
        in: body
        name: kermesse
        required: true
        schema:
          $ref: '#/definitions/requests.KermeseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Groupe créé
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Voyage non trouvé
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Créé une kermesse
      tags:
      - Kermesse
...
  /kermesses:
    get:
      consumes:
      - application/json
      description: Fetches kermesses for admin, organizers, and participants
      parameters:
      - default: Bearer Add access token here
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of kermesses
          schema:
            items:
              $ref: '#/definitions/models.Kermesse'
            type: array
        "401":
          description: User not logged
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Get all Kermesses based on user role
      tags:
      - Kermesse
  /kermesses/{id}:
    get:
      consumes:
      - application/json
      description: Fetch a kermesse by its ID if the user has permission
      parameters:
      - default: Bearer Add access token here
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Kermesse ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Kermesse found
          schema:
            $ref: '#/definitions/models.Kermesse'
        "401":
          description: User not logged
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Kermesse not found
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Get a Kermesse by its ID
      tags:
      - Kermesse
  /kermesses/{id}/add-stands:
    post:
      consumes:
      - application/json
      description: Permet d'ajouter des users (partcicpants ou organisateurs)
      parameters:
      - default: Bearer Add access token here
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Kermesse ID
        in: path
        name: id
        required: true
        type: integer
      - description: Données du groupe
        in: body
        name: kermesse
        required: true
        schema:
          $ref: '#/definitions/requests.AddStandRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Stand ajouté
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: kermesse non trouvé
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Ajouter des stands à la kermesse
      tags:
      - Kermesse
  /kermesses/{id}/add-users:
    post:
      consumes:
      - application/json
      description: Permet aux user de créé un groupe de groupeVoyage
      parameters:
      - default: Bearer Add access token here
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Kermesse ID
        in: path
        name: id
        required: true
        type: integer
      - description: Données du groupe
        in: body
        name: kermesse
        required: true
        schema:
          $ref: '#/definitions/requests.AddUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User ajouté(s)
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: kermesse non trouvé
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Ajouter des users à la kermesse
      tags:
      - Kermesse
  /kermesses/{id}/delete:
    delete:
      consumes:
      - application/json
      description: Allows an admin or the creator to update a Kermesse
      parameters:
      - default: Bearer Add access token here
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Kermesse ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Kermesse delete
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: User not logged
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Kermesse not found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Delete a Kermesse
      tags:
      - Kermesse
  /kermesses/{id}/update:
    put:
      consumes:
      - application/json
      description: Allows an admin or the creator to update a Kermesse
      parameters:
      - default: Bearer Add access token here
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Kermesse ID
        in: path
        name: id
        required: true
        type: integer
      - description: Kermesse data
        in: body
        name: kermesse
        required: true
        schema:
          $ref: '#/definitions/models.Kermesse'
      produces:
      - application/json
      responses:
        "200":
          description: Kermesse updated
          schema:
            $ref: '#/definitions/models.Kermesse'
        "401":
          description: User not logged
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Kermesse not found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Update a Kermesse
      tags:
      - Kermesse
...
swagger: "2.0"